% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_impute.R
\name{mean_impute}
\alias{mean_impute}
\title{Impute the median for continuous variables, and the mode for categorical variables.}
\usage{
mean_impute(data = NULL)
}
\arguments{
\item{data}{dataframe of the data}
}
\value{
Returns a plot showing the OR of the other variables for the missingness of the variable of interest
}
\description{
Impute the median for continuous variables, and the mode for categorical variables.
}
