% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Plot multiple ggplot objects}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot object}

\item{plotlist}{list of ggplot objects. Either plotlist or ... is specified.}

\item{file}{...}

\item{cols}{number of columns}

\item{layout}{matrix describing how the plots should be ordered}
}
\value{
Returns a plot containing all ggplot objects
}
\description{
Plot multiple ggplot objects
}
