% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prepost.R
\name{plot_prepost}
\alias{plot_prepost}
\title{Plots the distribution of all variables in the completed versus the original dataset.}
\usage{
plot_prepost(data = NULL, orig_data = NULL, numcols = 1)
}
\arguments{
\item{data}{dataframe of the completed data, can also be a multiple imputed dataset (mids), imputed by MICE}

\item{orig_data}{dataframe of the incompleted dataset}

\item{numcols}{number of columns in the final plot}
}
\value{
Returns a plot with multiple panels, for each variable of the dataset, comparing the distribution with the original data.
}
\description{
Plots the distribution of all variables in the completed versus the original dataset.
}
