% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relation_missing.R
\name{plot_relation_missing}
\alias{plot_relation_missing}
\title{Plot the relationship between the variables of the dataset and the
missingness of the variable of interest}
\usage{
plot_relation_missing(varbl = NULL, data = NULL)
}
\arguments{
\item{varbl}{variable of interest (character)}

\item{data}{dataframe of the data (including the variable of interest)}
}
\value{
Returns a plot showing the OR of the other variables for the missingness of the variable of interest
}
\description{
Plot the relationship between the variables of the dataset and the
missingness of the variable of interest
}
